refactor: move authentication components to libs/auth directory
