+++ src/libs/auth/components/SignOut.tsx
@@ -1,4 +1,4 @@
-import { authReactClient } from "src/utils/auth-client";
+import { authReactClient } from "../auth-client";
 
 export const SignOut = () => {
   return (
