+++ src/pages/index.astro
@@ -1,9 +1,8 @@
 ---
 import { CreateRoom } from "~libs/game-room/components/CreateRoom";
-import { Login } from "../components/Login";
-import { SignUp } from "../components/SignUp";
-import { SignOut } from "../components/SignOut";
-import type { ExpandedRoom } from "~libs/types/Room";
+import { Login } from "~libs/auth/components/Login";
+import { SignUp } from "~libs/auth/components/SignUp";
+import { SignOut } from "~libs/auth/components/SignOut";
 import { db } from "~libs/drizzle-orm/db";
 import { gameRoom } from "db/schemas/schema";
 import { eq } from "drizzle-orm";
